requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
supabase==2.10.0
pydantic==2.12.5
python-dotenv==1.0.1
httpx==0.27.2
anyio==4.12.1
starlette==0.38.6
postgrest==0.18.0
